Change data type of id

+++ models/tanaman_model.go
@@ -1,7 +1,7 @@
 package models
 
 type Tanaman struct {
-	Id           int64  `gorm:"primaryKey" json:"id"`
+	Id           int    `gorm:"primaryKey" json:"id"`
 	NamaTanaman  string `gorm:"type:varchar(250);unique" json:"nama_tanaman"`
 	JenisTanaman string `gorm:"type:varchar(250)" json:"jenis_tanaman"`
 	Deskripsi    string `gorm:"type:text" json:"deskripsi"`
